pyproject.toml
[tool.poetry]
name = "tulr"
version = "0.1.0"
description = "Simple url tracker"
authors = ["Pedro Silva <pedro.figs.silva@gmail.com>"]
license = "MIT"
packages = [
    { include = "turl" }
]

[tool.poetry.dependencies]
python = "^3.8"
uvloop = "^0.16.0"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
PyYAML = "^6.0"
confluent-kafka = "^1.8.2"
aiopg = "^1.3.3"
schema = "^0.7.5"
randomname = "^0.1.5"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
turl = "turl.__main__:entrypoint_turl"
turl_pg_sink = "turl.__main__:entrypoint_pg_sink"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
